+++ src/refiner/inference/providers/google.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from collections.abc import Mapping, Sequence
-from typing import Any
+from typing import Any, cast
 
 from refiner.inference._media import (
     base64_data,
@@ -14,7 +14,12 @@ from refiner.inference._message_conversion import (
     _provider_options,
 )
 from refiner.inference._schema import StructuredOutputSchema
-from refiner.inference.types import Message, ProviderOptions
+from refiner.inference._response import (
+    InferenceResponse,
+    _provider_metadata,
+    _text_from_content,
+)
+from refiner.inference.types import Message, ProviderOptions, ResponseContentPart
 
 PROVIDER_OPTIONS = {
     "apiClient",
@@ -226,4 +231,126 @@ def _apply_google_options(payload: dict[str, Any], options: Mapping[str, Any]) -
             payload[key] = options[key]
 
 
-__all__ = ["PROVIDER_OPTIONS", "build_payload"]
+def parse_response(
+    response_json: Mapping[str, Any],
+    *,
+    response_headers: Mapping[str, str] | None = None,
+) -> InferenceResponse:
+    candidates = response_json.get("candidates")
+    if not isinstance(candidates, Sequence) or not candidates:
+        raise RuntimeError("google generation response is missing candidates[0]")
+    candidate = candidates[0]
+    if not isinstance(candidate, Mapping):
+        raise RuntimeError("google generation response candidates[0] must be an object")
+    content = candidate.get("content")
+    if not isinstance(content, Mapping):
+        raise RuntimeError("google generation response is missing content")
+    parts = content.get("parts")
+    if not isinstance(parts, Sequence):
+        raise RuntimeError("google generation response is missing content.parts")
+    content_parts: list[ResponseContentPart] = []
+    for part in parts:
+        if not isinstance(part, Mapping):
+            continue
+        if isinstance(part.get("text"), str):
+            if part.get("thought") is True:
+                content_parts.append({"type": "reasoning", "text": part["text"]})
+            else:
+                content_parts.append({"type": "text", "text": part["text"]})
+        content_parts.extend(_google_generated_parts(part))
+    content_parts.extend(_google_grounding_sources(candidate))
+    text = _text_from_content(content_parts)
+    if not text:
+        raise RuntimeError("google generation response is missing textual content")
+    usage_metadata = response_json.get("usageMetadata")
+    usage = _google_usage(usage_metadata if isinstance(usage_metadata, Mapping) else {})
+    finish_reason = candidate.get("finishReason")
+    if finish_reason is not None and not isinstance(finish_reason, str):
+        finish_reason = str(finish_reason)
+    return InferenceResponse(
+        text=text,
+        finish_reason=finish_reason,
+        usage=usage,
+        response=response_json,
+        content=content_parts,
+        headers=dict(response_headers or {}),
+        provider_metadata=_provider_metadata("google", response_json, candidate),
+    )
+
+
+def _google_generated_parts(part: Mapping[str, Any]) -> list[ResponseContentPart]:
+    inline_data = part.get("inlineData") or part.get("inline_data")
+    if isinstance(inline_data, Mapping):
+        media_type = inline_data.get("mimeType") or inline_data.get("mime_type")
+        data = inline_data.get("data")
+        top_level = media_type.split("/", 1)[0] if isinstance(media_type, str) else ""
+        result: dict[str, Any] = {
+            "type": "image" if top_level == "image" else "file",
+            "providerMetadata": {"google": dict(part)},
+        }
+        if isinstance(media_type, str):
+            result["mediaType"] = media_type
+        if isinstance(data, str):
+            result["data"] = data
+        return [cast(ResponseContentPart, result)]
+    file_data = part.get("fileData") or part.get("file_data")
+    if isinstance(file_data, Mapping):
+        media_type = file_data.get("mimeType") or file_data.get("mime_type")
+        url = file_data.get("fileUri") or file_data.get("file_uri")
+        top_level = media_type.split("/", 1)[0] if isinstance(media_type, str) else ""
+        result = {
+            "type": "image" if top_level == "image" else "file",
+            "providerMetadata": {"google": dict(part)},
+        }
+        if isinstance(media_type, str):
+            result["mediaType"] = media_type
+        if isinstance(url, str):
+            result["url"] = url
+        return [cast(ResponseContentPart, result)]
+    return []
+
+
+def _google_grounding_sources(
+    candidate: Mapping[str, Any],
+) -> list[ResponseContentPart]:
+    grounding = candidate.get("groundingMetadata")
+    if not isinstance(grounding, Mapping):
+        return []
+    chunks = grounding.get("groundingChunks")
+    if not isinstance(chunks, Sequence) or isinstance(chunks, str):
+        return []
+    sources: list[ResponseContentPart] = []
+    for chunk in chunks:
+        if not isinstance(chunk, Mapping):
+            continue
+        web = chunk.get("web")
+        if not isinstance(web, Mapping):
+            continue
+        url = web.get("uri")
+        if not isinstance(url, str):
+            continue
+        source: dict[str, Any] = {
+            "type": "source",
+            "sourceType": "url",
+            "url": url,
+            "providerMetadata": {"google": dict(chunk)},
+        }
+        title = web.get("title")
+        if isinstance(title, str):
+            source["title"] = title
+        sources.append(cast(ResponseContentPart, source))
+    return sources
+
+
+def _google_usage(usage_metadata: Mapping[str, Any]) -> Mapping[str, Any]:
+    usage: dict[str, Any] = {}
+    if "promptTokenCount" in usage_metadata:
+        usage["prompt_tokens"] = usage_metadata["promptTokenCount"]
+    if "candidatesTokenCount" in usage_metadata:
+        usage["completion_tokens"] = usage_metadata["candidatesTokenCount"]
+    if "totalTokenCount" in usage_metadata:
+        usage["total_tokens"] = usage_metadata["totalTokenCount"]
+    return usage
+
+
+__all__ = ["PROVIDER_OPTIONS", "build_payload", "parse_response"]
